% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_datasets}
\alias{get_datasets}
\title{Get Datasets Metadata}
\usage{
get_datasets(query)
}
\arguments{
\item{query}{List with optional \code{ds} element for dataset IDs}
}
\value{
Data frame with dataset metadata
}
\description{
Retrieves dataset metadata.
}
\keyword{internal}
