% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_datasets_index}
\alias{get_datasets_index}
\title{Get Datasets Index}
\usage{
get_datasets_index()
}
\value{
Data frame with dataset IDs and sizes
}
\description{
Retrieves dataset IDs and sizes.
}
\keyword{internal}
