% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_degs}
\alias{get_degs}
\title{Get Differentially Expressed Genes for Datasets}
\usage{
get_degs(query)
}
\arguments{
\item{query}{List with \code{ds} element containing dataset IDs}
}
\value{
Data frame with DEG information
}
\description{
Retrieves DEG information for datasets.
}
\keyword{internal}
