% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_genes}
\alias{get_genes}
\title{Get Genes for Datasets}
\usage{
get_genes(query)
}
\arguments{
\item{query}{List with \code{ds} element containing dataset IDs}
}
\value{
Data frame with gene information
}
\description{
Retrieves gene lists for datasets.
}
\keyword{internal}
