% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_publications}
\alias{get_publications}
\title{Get Publications Metadata}
\usage{
get_publications(query)
}
\arguments{
\item{query}{List with optional \code{pub} element for publication IDs}
}
\value{
Data frame with publication metadata
}
\description{
Retrieves publication metadata.
}
\keyword{internal}
