% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{load_ds}
\alias{load_ds}
\title{Load Datasets}
\usage{
load_ds(query)
}
\arguments{
\item{query}{List containing \code{ds} element with dataset IDs to load}
}
\value{
Character vector of currently loaded dataset IDs
}
\description{
Loads Seurat objects from QS2 files and caches them in memory.
}
\keyword{internal}
