% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumb.R
\name{msc_plumb}
\alias{msc_plumb}
\title{Launch the Plumber2 API}
\usage{
msc_plumb(port = plumber2::get_opts("port", 8080L), ...)
}
\arguments{
\item{port}{Port to run the API on. Default: 8080}

\item{...}{Options passed to \code{plumber2::api_run()}}
}
\value{
A running Plumber API
}
\description{
Starts the Plumber2 API server for the MultiSC-Viewer daemon.
}
