% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{save_plot}
\alias{save_plot}
\title{Save Plot to File}
\usage{
save_plot(plot_object, path, ...)
}
\arguments{
\item{plot_object}{A ggplot object to save}

\item{path}{Destination file path for the PNG output}

\item{...}{Additional arguments passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}
}
\description{
Saves a ggplot object as PNG.
}
\keyword{internal}
