% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{unload_ds}
\alias{unload_ds}
\title{Unload Datasets}
\usage{
unload_ds(query)
}
\arguments{
\item{query}{List containing optional \code{ds} element with dataset IDs to unload}
}
\value{
Character vector of remaining loaded dataset IDs
}
\description{
Unloads specified datasets. If none specified, unloads all datasets.
}
\keyword{internal}
